Tag ReactOS 0.3.1 release (2/2).

+++ reactos/lib/rtl/largeint.c
@@ -75,7 +75,7 @@ RtlEnlargedUnsignedDivide (
 )
 {
    if (Remainder)
-      *Remainder = (ULONG)(Dividend.QuadPart % Divisor);
+      *Remainder = Dividend.QuadPart % Divisor;
 
    return (ULONG)(Dividend.QuadPart / Divisor);
 }
@@ -128,7 +128,7 @@ RtlExtendedLargeIntegerDivide (
    LARGE_INTEGER RC;
 
    if (Remainder)
-      *Remainder = (ULONG)(Dividend.QuadPart % Divisor);
+      *Remainder = Dividend.QuadPart % Divisor;
 
    RC.QuadPart = Dividend.QuadPart / Divisor;
